Append books module
