+++ modules/addbooks.js
@@ -1,0 +1,19 @@
+import {list, books} from '../modules/index.js';
+
+export function appendBooksToList() {
+  list.innerHTML = '';
+  books.bookList.forEach((book, index) => {
+    const li = document.createElement('li');
+
+    if (index % 2 === 1) {
+      li.classList = 'bg-white';
+    }
+
+    li.innerHTML = `
+    <p>"<span>${book.title}</span>" </p>
+    <p>by ${book.author}</p>
+    <button class="remove">Remove</button>
+    `;
+    list.appendChild(li);
+  });
+}
