+++ modules/index.js
@@ -1,6 +1,7 @@
 import { addToLocalStorage, getFromLocalStorage } from './localstorage.js';
+import {appendBooksToList} from './addbooks.js';
 
-const list = document.querySelector('ul');
+export const list = document.querySelector('ul');
 const form = document.querySelector('form');
 const title = document.querySelector('#title');
 const author = document.querySelector('#author');
@@ -42,25 +43,6 @@ function checkIfEmpty() {
   }
 }
 
-
-
-function appendBooksToList() {
-  list.innerHTML = '';
-  books.bookList.forEach((book, index) => {
-    const li = document.createElement('li');
-
-    if (index % 2 === 1) {
-      li.classList = 'bg-white';
-    }
-
-    li.innerHTML = `
-    <p>"<span>${book.title}</span>" </p>
-    <p>by ${book.author}</p>
-    <button class="remove">Remove</button>
-    `;
-    list.appendChild(li);
-  });
-}
 
 export function updateDomAndLocalStorage() {
   appendBooksToList();
@@ -131,78 +113,3 @@ nav.addEventListener('click', (event) => {
     contactTab.classList.add('red');
   }
 });
-
-setInterval(() => {
-  const today = new Date();
-  let month = '';
-  let day = '';
-  let hour = today.getHours();
-  let hourType = 'am';
-
-  if (hour > 12) {
-    hour -= 12;
-    hourType = 'pm';
-  } else if (hour === 12) {
-    hourType = 'pm';
-  }
-
-  switch (today.getDate()) {
-    case 1:
-      day = 'st';
-      break;
-    case 2:
-      day = 'nd';
-      break;
-    case 3:
-      day = 'rd';
-      break;
-    default:
-      day = 'th';
-  }
-
-  switch (today.getMonth()) {
-    case 0:
-      month = 'January';
-      break;
-    case 1:
-      month = 'February';
-      break;
-    case 2:
-      month = 'March';
-      break;
-    case 3:
-      month = 'April';
-      break;
-    case 4:
-      month = 'May';
-      break;
-    case 5:
-      month = 'June';
-      break;
-    case 6:
-      month = 'July';
-      break;
-    case 7:
-      month = 'August';
-      break;
-    case 8:
-      month = 'September';
-      break;
-    case 9:
-      month = 'October';
-      break;
-    case 10:
-      month = 'November';
-      break;
-    case 11:
-      month = 'December';
-      break;
-    default:
-      month = 'Do not know';
-  }
-
-  const date = `${month} ${today.getDate()}${day} ${today.getFullYear()},`;
-  const time = `${hour}:${today.getMinutes()}:${today.getSeconds()} ${hourType}`;
-  const dateTime = `${date} ${time}`;
-  document.querySelector('.time-and-date').innerHTML = dateTime;
-}, 1000);
